Update formatting of the views module (#24)
